Move stuff into views directory
