+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 // Load in other views here
-import Specials from './Specials';
+import Specials from './views/Specials';
 
 
 // Very simple layout for now; Eventually need to set up navigation
